version 2021.04.14
modification in GevSeq.py for the place of the line named : # uncomment for preloaded root files
It have to be placed before the :
"os.chdir(webFolder) # going into finalFolder" order in order to copy all data into the final folder.
If not placed before the order, the data are copyed locally.